Completed Practical 8.
In this practical i have implemented a student management system, in that the teacher can deleted and update the student list and we can add student data to the list and we can find the student from the list.

+++ Practical8/Practical8/Practical8/IStudent.cs
@@ -13,6 +13,6 @@ namespace Practical8
         string LastName { get; set; }
         int Age { get; set; }
         string Email { get; set; }
-        int PhoneNumber { get; set; }
+        long PhoneNumber { get; set; }
     }
 }
